# docs/openapi/declaration.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RÉCOR Declaration Service",
    "description": "REST surface of the RÉCOR Declaration service. Accepts beneficial-ownership declarations, returns signed receipts, relays verification outcomes from the Verification Engine, and exposes operator endpoints under `/v1/internal/*`. Endpoints under the `internal` tag are not for general consumer use.",
    "contact": {
      "name": "RÉCOR domain team",
      "email": "noreply@recor.cm"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local dev"
    },
    {
      "url": "https://api.recor.cm",
      "description": "Production (TBD)"
    }
  ],
  "paths": {
...
    "/v1/internal/verification-outcomes": {
      "post": {
        "tags": [
          "internal"
        ],
        "operationId": "handleVerificationOutcome",
        "requestBody": {
          "description": "Slim outbox envelope from the Verification Engine; payload conforms to VerificationOutcomeRequest.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationOutcomeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Replay of an already-recorded outcome (no-op)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationOutcomeResponse"
                }
              }
            }
          },
          "201": {
            "description": "New verification outcome recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationOutcomeResponse"
                }
              }
            }
          },
          "202": {
            "description": "Accepted but ignored (unknown event_type)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationOutcomeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed envelope or payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or bad HMAC signature",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Referenced declaration does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflicting outcome for the same case_id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Backend failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Writeback endpoint disabled (no HMAC secret configured)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "hmacSignature": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "DeclarationId": {
        "type": "string",
        "format": "uuid",
        "description": "Stable identifier for a declaration. UUIDv7 — time-sortable, so\nnatural ordering matches submission order.",
        "example": "0192f1d4-1e0a-7c4b-9b1e-3d4f5a6b7c8d"
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Inner body of an `ErrorEnvelope`. `kind` is a stable machine-friendly\nclassifier — clients SHOULD switch on it. `message` is a human\ndescription for logs; clients SHOULD NOT parse it.",
        "required": [
          "kind",
          "message"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "description": "Machine-friendly classifier (e.g. `bad_request`, `not_found`,\n`conflict`, `forbidden`, `authentication_required`,\n`idempotency_conflict`, `bad_attestation`, `internal`,\n`optimistic_concurrency_conflict`, `admin_disabled`, `not_admin`,\n`missing_signature`, `bad_signature`, `writeback_disabled`,\n`dlq_row_not_found`, `malformed_envelope`, `malformed_payload`)."
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "description": "Standard error response body. Every non-2xx response is shaped like\n`{ \"error\": { \"kind\": \"<machine>\", \"message\": \"<human>\" } }`.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "VerificationLane": {
        "type": "string",
        "description": "Lane decision returned by the Verification Engine for a declaration.\nMatches the verification engine's LaneDecision enum byte-for-byte over\nthe wire (snake_case serialisation).",
        "enum": [
          "green",
          "yellow",
          "red"
        ]
      },
      "VerificationOutcomeRequest": {
        "type": "object",
        "description": "Inbound envelope on POST /v1/internal/verification-outcomes.\n\nField names + types MUST match the verification engine's outbox\npayload exactly. See\nservices/verification-engine/src/infrastructure/postgres.rs writeback\npayload construction.",
        "required": [
          "case_id",
          "declaration_id",
          "lane",
          "fused_authenticity_belief",
          "fused_authenticity_plausibility",
          "fused_risk_belief",
          "completed_at"
        ],
        "properties": {
          "case_id": {
            "type": "string",
            "format": "uuid"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time"
          },
          "declaration_id": {
            "$ref": "#/components/schemas/DeclarationId"
          },
          "fused_authenticity_belief": {
            "type": "number",
            "format": "double",
            "description": "Dempster-Shafer belief in payload authenticity, in [0.0, 1.0].",
            "maximum": 1,
            "minimum": 0
          },
          "fused_authenticity_plausibility": {
            "type": "number",
            "format": "double",
            "description": "Dempster-Shafer plausibility of payload authenticity, in [0.0, 1.0].",
            "maximum": 1,
            "minimum": 0
          },
          "fused_risk_belief": {
            "type": "number",
            "format": "double",
            "description": "Dempster-Shafer belief that this declaration carries risk, in [0.0, 1.0].",
            "maximum": 1,
            "minimum": 0
          },
          "lane": {
            "$ref": "#/components/schemas/VerificationLane"
          }
        }
      },
      "VerificationOutcomeResponse": {
        "type": "object",
        "required": [
          "declaration_id",
          "verification_case_id",
          "lane",
          "recorded_new_event"
        ],
        "properties": {
          "declaration_id": {
            "$ref": "#/components/schemas/DeclarationId"
          },
          "lane": {
            "$ref": "#/components/schemas/VerificationLane"
          },
          "recorded_new_event": {
            "type": "boolean",
            "description": "`true` when the outcome produced a new event; `false` on a\nno-op replay (the verification engine relay is at-least-once,\nso this distinguishes new outcomes from retries)."
          },
          "verification_case_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OIDC-issued JWT. The service verifies the signature against the configured issuer's JWKS, plus `iss`/`aud`/`exp`/`nbf` claims. HMAC algorithms are refused (algorithm-confusion mitigation)."
      },
      "devPrincipalHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Recor-Dev-Principal",
        "description": "Dev-only principal shortcut. Refused outside ENVIRONMENT=dev. Production must always use `bearerAuth`."
      },
      "hmacSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RECOR-Signature",
        "description": "HMAC-SHA256 over the raw request body, hex-encoded, using the shared writeback secret. Constant-time verification; rotation supported via a transient old-secret window."
      }
    }
  },
...
}